+++ extensions/wechat/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Wechat Changelog
 
+## [Copy url action] - {PR_MERGE_DATE}
+
+- Add action to copy quick access url for reuse
+- Update logic to launch WeChat
+
 ## [Initial Version] - 2022-09-19